Add support for Google FlatBuffers
Since I plan on having most (if not all) my data serialized and transmittable to a web client, it makes sense to include a serialization library. It also means that serialization/deserialization code is free and in sync between client and server. I have seen a lot of security bugs about missing bounds checks and parceling mismatches. Yay for canonical sources!