Rewrite WalkingRobotClip for workspace clipping.
I rewrote a good chunk of it. Now it's easier to understand and work with.
Future plans are to plan steps only that are inside the workspace. This new
architecture makes it easier for plans to actually move the foot land point.
Also as a result of the cleaner code, some hidden bugs were squashed. One of
them turning out to cause a not-quite-bug, but perhaps odd behavior.