Add motor control for the pololu mini maestro.
I had to do some makefile tweaks to get the wiringPi library up and then
learned I could be more portable without it. I found out a way to select a
hardware configuration now. My fault was having a "null" implementation of the
motor control interface; it's way easier to just have a null pointer to that
object when the implementation doesn't exist. This allows single-line options
for specifiying a compile-time option.